Add factory address to event

+++ test/PoolFactory.t.sol
@@ -40,26 +40,18 @@ contract PoolFactoryTest is Test {
     function setUp() public {
         mockWvtru = new Token_ERC20("TKN", "TKN", 18);
         factory = new PoolFactory(
-            priceContract,
-            _creationFeeWallet,
-            _bankWallet,
-            _airDropWallet,
-            _feeWallet,
-            address(mockWvtru)
+            priceContract, _creationFeeWallet, _bankWallet, _airDropWallet, _feeWallet, address(mockWvtru)
         );
     }
 
     function test_CalculateTokensForWVTRU() public view {
-        (uint256 amount, ) = factory.tokensForWvtru(1 ether, TOKEN_SUPPLY);
+        (uint256 amount,) = factory.tokensForWvtru(1 ether, TOKEN_SUPPLY);
 
         assertEq(amount, 194001587685638243627);
     }
 
     function test_CalculateWVTRUForTokens() public view {
-        (uint amount, ) = factory.wvtruForTokens(
-            195960807838432313537,
-            TOKEN_SUPPLY
-        );
+        (uint256 amount,) = factory.wvtruForTokens(195960807838432313537, TOKEN_SUPPLY);
 
         assertApproxEqAbs(amount, 0.99 ether, 1);
     }
@@ -67,29 +59,14 @@ contract PoolFactoryTest is Test {
     function test_CreatePool() public {
         mockWvtru.mint(address(this), priceContract);
         IERC20(address(mockWvtru)).approve(address(factory), priceContract);
-        (address poolAddress, address tokenAddress) = factory
-            .createPoolWithToken(
-                "TKN",
-                "TKN",
-                "url",
-                TOKEN_SUPPLY,
-                priceContract
-            );
+        (address poolAddress, address tokenAddress) =
+            factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, priceContract);
 
         IERC20 token = IERC20(tokenAddress);
 
-        assertEq(
-            token.balanceOf(_bankWallet),
-            (TOKEN_SUPPLY / 100_0000) * 4211
-        );
-        assertEq(
-            token.balanceOf(_airDropWallet),
-            (TOKEN_SUPPLY / 100_0000) * 1_5789
-        );
-        assertEq(
-            token.balanceOf(poolAddress),
-            (TOKEN_SUPPLY / 100_0000) * 98_0000
-        );
+        assertEq(token.balanceOf(_bankWallet), (TOKEN_SUPPLY / 100_0000) * 4211);
+        assertEq(token.balanceOf(_airDropWallet), (TOKEN_SUPPLY / 100_0000) * 1_5789);
+        assertEq(token.balanceOf(poolAddress), (TOKEN_SUPPLY / 100_0000) * 98_0000);
         assertTrue(poolAddress != address(0));
         assertTrue(tokenAddress != address(0));
     }
@@ -99,19 +76,10 @@ contract PoolFactoryTest is Test {
         uint256 amount = priceContract + wvtruToBuy;
         mockWvtru.mint(address(this), amount);
         IERC20(address(mockWvtru)).approve(address(factory), amount);
-        (, address tokenAddress) = factory.createPoolWithToken(
-            "TKN",
-            "TKN",
-            "url",
-            TOKEN_SUPPLY,
-            amount
-        );
+        (, address tokenAddress) = factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
 
         uint256 balance = IERC20(tokenAddress).balanceOf(address(this));
-        (uint256 amountToken, ) = factory.tokensForWvtru(
-            wvtruToBuy,
-            TOKEN_SUPPLY
-        );
+        (uint256 amountToken,) = factory.tokensForWvtru(wvtruToBuy, TOKEN_SUPPLY);
 
         assertEq(balance, amountToken);
     }
@@ -121,20 +89,17 @@ contract PoolFactoryTest is Test {
         uint256 amount = priceContract;
         mockWvtru.mint(address(this), amount + amountToBuy);
         IERC20(address(mockWvtru)).approve(address(factory), amount);
-        (address poolAddress, address tokenAddress) = factory
-            .createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
+        (address poolAddress, address tokenAddress) =
+            factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
 
         LiquidityPool pool = LiquidityPool(payable(poolAddress));
         Token token = Token(tokenAddress);
 
         IERC20(address(mockWvtru)).approve(address(poolAddress), amountToBuy);
 
-        (uint256 expectedBalance, ) = factory.tokensForWvtru(
-            amountToBuy,
-            TOKEN_SUPPLY
-        );
+        (uint256 expectedBalance,) = factory.tokensForWvtru(amountToBuy, TOKEN_SUPPLY);
 
-        uint balanceFromPool = pool.calcOutputToken(amountToBuy);
+        uint256 balanceFromPool = pool.calcOutputToken(amountToBuy);
 
         pool.buyToken(amountToBuy);
         uint256 balance = token.balanceOf(address(this));
@@ -148,8 +113,8 @@ contract PoolFactoryTest is Test {
         uint256 amount = priceContract;
         mockWvtru.mint(address(this), amount + amountToBuy);
         IERC20(address(mockWvtru)).approve(address(factory), amount);
-        (address poolAddress, address tokenAddress) = factory
-            .createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
+        (address poolAddress, address tokenAddress) =
+            factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
 
         LiquidityPool pool = LiquidityPool(payable(poolAddress));
         Token token = Token(tokenAddress);
@@ -160,14 +125,11 @@ contract PoolFactoryTest is Test {
 
         uint256 tokenBalance = token.balanceOf(address(this));
         IERC20(tokenAddress).approve(address(poolAddress), tokenBalance);
-        uint balanceFromPool = pool.calcOutputVtru(tokenBalance);
+        uint256 balanceFromPool = pool.calcOutputVtru(tokenBalance);
 
         pool.sellToken(tokenBalance);
 
-        (uint256 expectedBalance, ) = factory.wvtruForTokens(
-            tokenBalance,
-            TOKEN_SUPPLY
-        );
+        (uint256 expectedBalance,) = factory.wvtruForTokens(tokenBalance, TOKEN_SUPPLY);
 
         uint256 balance = IERC20(address(mockWvtru)).balanceOf(address(this));
 
@@ -178,20 +140,13 @@ contract PoolFactoryTest is Test {
     function test_InitialMarketCap() public {
         mockWvtru.mint(address(this), priceContract);
         IERC20(address(mockWvtru)).approve(address(factory), priceContract);
-        (address poolAddress, ) = factory.createPoolWithToken(
-            "TKN",
-            "TKN",
-            "url",
-            TOKEN_SUPPLY,
-            priceContract
-        );
+        (address poolAddress,) = factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, priceContract);
 
         LiquidityPool pool = LiquidityPool(payable(poolAddress));
 
         uint256 priceForToken = pool.calcOutputVtru(1 ether);
 
-        uint256 totalPriceForTokens = (pool.realTokenBalance() / 1 ether) *
-            priceForToken;
+        uint256 totalPriceForTokens = (pool.realTokenBalance() / 1 ether) * priceForToken;
 
         console.log(totalPriceForTokens);
         assertApproxEqAbs(totalPriceForTokens, 5000 ether, 100 ether);
@@ -202,13 +157,7 @@ contract PoolFactoryTest is Test {
         uint256 amount = priceContract;
         mockWvtru.mint(address(this), amount + amountToBuy);
         IERC20(address(mockWvtru)).approve(address(factory), amount);
-        (address poolAddress, ) = factory.createPoolWithToken(
-            "TKN",
-            "TKN",
-            "url",
-            TOKEN_SUPPLY,
-            amount
-        );
+        (address poolAddress,) = factory.createPoolWithToken("TKN", "TKN", "url", TOKEN_SUPPLY, amount);
 
         LiquidityPool pool = LiquidityPool(payable(poolAddress));
 
@@ -218,8 +167,7 @@ contract PoolFactoryTest is Test {
 
         uint256 priceForToken = pool.calcOutputVtru(1 ether);
 
-        uint256 totalPriceForTokens = (pool.realTokenBalance() / 1 ether) *
-            priceForToken;
+        uint256 totalPriceForTokens = (pool.realTokenBalance() / 1 ether) * priceForToken;
 
         console.log(totalPriceForTokens);
